refactor: separate getProductsListFromDB
